requirements.txt
sqlalchemy
aiosqlite
aiomysql
redis
python-dotenv
python3-nmap
sendgrid
vulners
django
numpy
django_bootstrap5
APScheduler
mysqlclient